chore(virtctl): Move migrate command to own file
